Extract dynamodb database code into package
* Extract all dynamodb database code into separate package.
* Remove other inventory store types since they're no longer used
* Update all handlers to use the new package
