+++ handlers/ipam-ip/main_test.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	dynamodbtest "github.com/PolarGeospatialCenter/dockertest/pkg/dynamodb"
-	"github.com/PolarGeospatialCenter/inventory/pkg/inventory"
+	"github.com/PolarGeospatialCenter/inventory/pkg/inventory/dynamodbclient"
 	"github.com/PolarGeospatialCenter/inventory/pkg/inventory/types"
 	inventorytypes "github.com/PolarGeospatialCenter/inventory/pkg/inventory/types"
 	"github.com/PolarGeospatialCenter/inventory/pkg/lambdautils"
@@ -31,7 +31,7 @@ func runTest(t *testing.T, h testHandler) {
 	defer dbInstance.Stop(ctx)
 
 	db := dynamodb.New(session.New(dbInstance.Config()))
-	inv := inventory.NewDynamoDBStore(db, nil)
+	inv := dynamodbclient.NewDynamoDBStore(db, nil)
 
 	err = inv.InitializeTables()
 	if err != nil {
