--- pkg/inventory/dynamodbstore/db.go
@@ -1,1 +1,0 @@
-package dynamodbstore
