--- pkg/inventory/dynamodbstore/ipam_test.go
@@ -1,1 +1,0 @@
-package dynamodbstore
